Moved YYFILL points calculation to the earlier stage of DFA construction.
No serious changes intended (mostly cleanup and comments).

The underlying algorithm for finding strongly connected components
(SCC) remains the same: it's a slightly modified Tarjan's algorithm.

We now mark non-YYFILL states by setting YYFILL argument to zero,
which is only logical: why would anyone call YYFILL to provide zero
characters. In fact, re2c didn't generate 'YYFILL(0)' call itself,
but some remnants of YYFILL did remain (which caused changes in tests).

+++ re2c/src/codegen/input_api.cc
@@ -157,7 +157,7 @@ std::string InputAPI::expr_lessthan_one () const
 		: expr_lessthan (1);
 }
 
-std::string InputAPI::expr_lessthan (uint32_t n) const
+std::string InputAPI::expr_lessthan (size_t n) const
 {
 	std::ostringstream s;
 	switch (type_)
